modelos

+++ entity_satellite.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Satellite struct {
 	gorm.Model
-	Name     string `json:"name"`
-	Distance int32  `json:"distance"`
-	Message  string `json:"message"`
+	Name     string  `json:"name"`
+	Distance float64 `json:"distance"`
+	Message  string  `json:"message"`
 }
